Avoid broken pipe errors when buffer is closed
When writing to the buffer, gpg can be already closed. One example when this happens is when invoking gpg for encryption but the key does not exist. gpg will exit immediately with exit code 2, but the copy_data thread will still try to write the file to be encrypted to stdin.

Note: This fix does not fully avoid all cases of race conditions, as the check is not truly atomic, but it improves the current situation.